Add workflow to automatize the testing process
The workflow `test-module.yml` will setup a testing environment and
execute the tests present in the `tests/` directory using the
`test-module.sh` script.
The workflow will be invoked every time that the workflow `Publish images`
had run with success, and can be also invoked manually using the
`workflow_dispatch` event.